+++ src/HttpClientRequestInterface.php
@@ -28,6 +28,15 @@ interface HttpClientRequestInterface
     public function addHeader(string $header): void;
 
     /**
+     * Returns the certificate authority (CA) certificate path or null if not set.
+     *
+     * @since 1.1.0
+     *
+     * @return FilePathInterface|null The certificate authority (CA) certificate path or null if not set.
+     */
+    public function getCACertificate(): ?FilePathInterface;
+
+    /**
      * Returns the files to upload.
      *
      * @since 1.0.0
@@ -80,6 +89,15 @@ interface HttpClientRequestInterface
      * @return UrlInterface The url.
      */
     public function getUrl(): UrlInterface;
+
+    /**
+     * Sets the certificate authority (CA) certificate path.
+     *
+     * @since 1.1.0
+     *
+     * @param FilePathInterface $caCertificate The certificate authority (CA) certificate path.
+     */
+    public function setCACertificate(FilePathInterface $caCertificate): void;
 
     /**
      * Sets a file to upload.