+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## Unreleased
+- Added getCACertificate and setCACertificate methods to HttpClientRequest.
 
 ## 1.0.0 - 2019-01-24
 - First stable revision.